(Mostly) Finished Exercises 7.6
Broke my teeth on trying to set up RSpec tests for testing for presence of errors in the error message.  Attempted to use @user.errors.count on an invalid submission to test if any errors thrown.  Got a class error problem with this, so I have more research to do.  Most of the other exercises complete, though not pretty.

Don't merge this with the master branch - all tests pass, but it's not clean...