Finish Section 9.3.1
Completed 9.3.1, but still getting an error for the test:
"Authentication authorization for non-signed-in users in the Users controller visiting the user index" in the authentication_pages_spec file.

Went through code from Ch 8 on, can't find an error.

Note, reverted the sessions/new.html.erb file to the original format using 'form_for' - I had refactored this to the 'form_tag' version in the exercises for Chapter 8.